feat: misc type cleanup / reorg
