+++ apps/web/src/app/(networks)/(evm)/perps/_ui/account-management/deposit-dialog/any-token-deposit.tsx
@@ -47,9 +47,7 @@ import type { AnyTokenDepositOption } from './deposit-dialog'
 import { USDCOptions, getMinDepositAmount, getUSDCArgs } from './usdc-options'
 
 type TokenToSwap = {
-  currency: EvmCurrency<{
-    approved: boolean
-  }>
+  currency: EvmCurrency
   balance: Amount<EvmCurrency> | undefined
   price: Fraction | undefined
   usd: string | undefined
