+++ pnpm-workspace.yaml
@@ -21,7 +21,7 @@ catalogs:
     "@wagmi/core": 3.2.2
     "viem": 2.44.4
     "@solana/addresses": 6.0.1
-    "sushi": 6.6.7
+    "sushi": 6.6.11
 
   react:
     "next": 16.2.6
